Make `Tracks` a type alias for Vec<Track>
I thought about this for a couple of hours, even considered
a #[repr(transparent)] newtype struct with some transmute/unsafe
code to cast the references from &Vec<Track> to &Tracks.

In the end I came to the conclusion that a type alias is the only
non-unsafe and reasonable design, which also elides unnecessary copies
in RocketClient::save_tracks and Player::new, but gives a type-level hint
about the intended API usage.